pyproject.toml

[tool.poetry]
name = "nonebot-plugin-anywhere-llm"
version = "0.6.5"
authors = ["Zeta-qixi <1019289695@qq.com>"]
description = "Nonebot2 plugin for using llm"
readme = "README.md"
license = "MIT"
homepage = "https://github.com/Zeta-qixi/nonebot-plugin-anywhere-llm"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.4.1"
nonebot-adapter-onebot="^2.4.3"
aiosqlite="^0.21.0"
openai="^1.55.3"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"